[PTI-LIB] Allow User to Modify Global CMake Project Settings (#709)
Allow users to modify global project settings by first checking if
they are being modified on the command line. This will be helpful for some
users with stricter C++ compilers. On some versions of MSVC, setting a
newer C++ version is required to use some features. This should probably
be fixed in our code, but this will also improve user experience if
issues like these occur in the future.

Also, fix a "narrowing conversion" warning on some versions of MSVC.

Improve CMake slightly (refactor includes for less repetition).

+++ sdk/src/levelzero/ze_collector.h
@@ -525,7 +525,7 @@ class ZeCollector {
                                     ._queue_handle = queue_handle,
                                     ._device_handle = command.device,
                                     ._phase = phase,
-                                    ._return_code = return_code,
+                                    ._return_code = static_cast<uint32_t>(return_code),
                                     ._correlation_id = command.corr_id_,
                                     ._operation_count = 1,
                                     ._operation_details = op_details};
